nntp: refactor Account/Mailbox/Email private data
