+++ Makefile.autosetup
@@ -465,7 +465,7 @@ $(PWD)/ncrypt:
 # libnntp
 LIBNNTP=	libnntp.a
 LIBNNTPOBJS=	nntp/browse.o nntp/config.o nntp/complete.o nntp/newsrc.o \
-		nntp/nntp.o
+		nntp/nntp.o nntp/adata.o nntp/edata.o nntp/mdata.o
 CLEANFILES+=	$(LIBNNTP) $(LIBNNTPOBJS)
 ALLOBJS+=	$(LIBNNTPOBJS)
