+++ debug/graphviz.c
@@ -46,8 +46,10 @@
 #include "maildir/mdata.h"
 #include "maildir/private.h"
 #include "mutt_globals.h"
-#include "notmuch/mdata.h"
+#include "nntp/adata.h"
+#include "nntp/mdata.h"
 #include "notmuch/adata.h"
+#include "notmuch/mdata.h"
 #include "notmuch/private.h"
 #include "pop/adata.h"
 #include "pop/private.h"
