+++ index.c
@@ -87,6 +87,8 @@
 #endif
 #ifdef USE_NNTP
 #include "nntp/lib.h"
+#include "nntp/adata.h"
+#include "nntp/mdata.h"
 #endif
 #ifdef ENABLE_NLS
 #include <libintl.h>
