+++ main.c
@@ -87,6 +87,8 @@
 #endif
 #ifdef USE_NNTP
 #include "nntp/lib.h"
+#include "nntp/adata.h"
+#include "nntp/mdata.h"
 #endif
 #ifdef USE_AUTOCRYPT
 #include "autocrypt/lib.h"
