+++ nntp/newsrc.c
@@ -47,7 +47,10 @@
 #include "mutt.h"
 #include "lib.h"
 #include "bcache/lib.h"
+#include "adata.h"
+#include "edata.h"
 #include "format_flags.h"
+#include "mdata.h"
 #include "mutt_account.h"
 #include "mutt_globals.h"
 #include "mutt_logging.h"
