+++ nntp/private.h
@@ -58,18 +58,16 @@ extern bool          C_ShowNewNews;
 
 void                    nntp_acache_free       (struct NntpMboxData *mdata);
 int                     nntp_active_save_cache (struct NntpAccountData *adata);
-struct NntpAccountData *nntp_adata_new         (struct Connection *conn);
 int                     nntp_add_group         (char *line, void *data);
 void                    nntp_article_status    (struct Mailbox *m, struct Email *e, char *group, anum_t anum);
 void                    nntp_bcache_update     (struct NntpMboxData *mdata);
 int                     nntp_check_new_groups  (struct Mailbox *m, struct NntpAccountData *adata);
 void                    nntp_delete_group_cache(struct NntpMboxData *mdata);
-struct NntpEmailData *  nntp_edata_get         (struct Email *e);
 void                    nntp_group_unread_stat (struct NntpMboxData *mdata);
 void                    nntp_hash_destructor_t (int type, void *obj, intptr_t data);
-struct HeaderCache *        nntp_hcache_open       (struct NntpMboxData *mdata);
+void                    nntp_hashelem_free     (int type, void *obj, intptr_t data);
+struct HeaderCache *    nntp_hcache_open       (struct NntpMboxData *mdata);
 void                    nntp_hcache_update     (struct NntpMboxData *mdata, struct HeaderCache *hc);
-void                    nntp_mdata_free        (void **ptr);
 void                    nntp_newsrc_gen_entries(struct Mailbox *m);
 int                     nntp_open_connection   (struct NntpAccountData *adata);
