+++ send/send.c
@@ -70,6 +70,7 @@
 #ifdef USE_NNTP
 #include "nntp/lib.h"
 #include "mx.h"
+#include "nntp/mdata.h"
 #endif
 #ifdef MIXMASTER
 #include "remailer.h"
